Created POS component
xxxx

- Skeleton for the POS component
