+++ core/domain/unittest/CMakeLists.txt
@@ -5,12 +5,12 @@ set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -ftest-coverage")
 set (CMAKE_EXE_LINKER_FLAGS "-fprofile-arcs -ftest-coverage -lgcov --coverage ${CMAKE_EXE_LINKER_FLAGS}")
 
 set (SOURCES_UNDER_TEST
+    ${DOMAIN_DIRECTORY}/accounting/accountingcontroller.cpp
     ${DOMAIN_DIRECTORY}/customermgmt/customermgmtcontroller.cpp
     ${DOMAIN_DIRECTORY}/dashboard/dashboardcontroller.cpp
     ${DOMAIN_DIRECTORY}/employeemgmt/employeecontroller.cpp
     ${DOMAIN_DIRECTORY}/inventory/inventorycontroller.cpp
-    ${DOMAIN_DIRECTORY}/accounting/accountingcontroller.cpp
-    ${DOMAIN_DIRECTORY}/accounting/salecomputer.cpp
+    ${DOMAIN_DIRECTORY}/pos/salecomputer.cpp
     ${DOMAIN_DIRECTORY}/userlogin/logincontroller.cpp
 )
